Enum boa_ast::declaration::ImportKind

pub enum ImportKind {
    DefaultOrUnnamed,
    Namespaced {
        binding: Identifier,
    },
    Named {
        names: Box<[ImportSpecifier]>,
    },
}

The kind of import in an ImportDeclaration.

Variants

DefaultOrUnnamed

Default (import defaultName from "module-name") or unnamed (`import “module-name”).

Namespaced

Fields

binding: Identifier

Binding for the namespace created from the exports of the imported module.

Namespaced import (import * as name from "module-name").

Named

Fields

names: Box<[ImportSpecifier]>

List of the required exports of the imported module.

Import list (import { export1, export2 as alias2 } from "module-name").