Module boa_ast::declaration::variable

Variable related declarations.

Structs

TryFromVariableListError

The error returned by the VariableList::try_from function.

VarDeclaration

A var statement, also called VariableStatement in the spec.

Variable

Variable represents a variable declaration of some kind.

VariableList

List of variables in a variable declaration.

Enums

Binding

Binding represents either an individual binding or a binding pattern.

LexicalDeclaration

A lexical declaration defines variables that are scoped to the lexical environment of the variable declaration.

Constants

RECURSIVE_COUNT_Binding ğŸ”’

RECURSIVE_COUNT_LexicalDeclaration ğŸ”’

RECURSIVE_COUNT_VarDeclaration ğŸ”’

RECURSIVE_COUNT_Variable ğŸ”’

RECURSIVE_COUNT_VariableList ğŸ”’

Statics

__KEY ğŸ”’

__KEY ğŸ”’

__KEY ğŸ”’

__KEY ğŸ”’

__KEY ğŸ”’

Functions

__getit ğŸ”’ âš

__getit ğŸ”’ âš

__getit ğŸ”’ âš

__getit ğŸ”’ âš

__getit ğŸ”’ âš

__init ğŸ”’

__init ğŸ”’

__init ğŸ”’

__init ğŸ”’

__init ğŸ”’