Enum boa_ast::Punctuator

pub enum Punctuator {

    Add,
    And,
    Arrow,
    Assign,
    AssignAdd,
    AssignAnd,
    AssignBoolAnd,
    AssignBoolOr,
    AssignCoalesce,
    AssignDiv,
    AssignLeftSh,
    AssignMod,
    AssignMul,
    AssignOr,
    AssignPow,
    AssignRightSh,
    AssignSub,
    AssignURightSh,
    AssignXor,
    BoolAnd,
    BoolOr,
    CloseBlock,
    CloseBracket,
    CloseParen,
    Coalesce,
    Colon,
    Comma,
    Dec,
    Div,
    Dot,
    Eq,
    GreaterThan,
    GreaterThanOrEq,
    Inc,
    LeftSh,
    LessThan,
    LessThanOrEq,
    Mod,
    Mul,
    Neg,
    Not,
    NotEq,
    OpenBlock,
    OpenBracket,
    OpenParen,
    Optional,
    Or,
    Exp,
    Question,
    RightSh,
    Semicolon,
    Spread,
    StrictEq,
    StrictNotEq,
    Sub,
    URightSh,
    Xor,

}

All of the punctuators used in ECMAScript.

More information:

ECMAScript Reference

Variants

Add

+

And

&

Arrow

=>

Assign

=

AssignAdd

+=

AssignAnd

&=

AssignBoolAnd

&&=

AssignBoolOr

||=

AssignCoalesce

??=,

AssignDiv

/=

AssignLeftSh

<<=

AssignMod

%=

AssignMul

*=

AssignOr

|=

AssignPow

**=

AssignRightSh

>>=

AssignSub

-=

AssignURightSh

>>>=

AssignXor

^=

BoolAnd

&&

BoolOr

||

CloseBlock

}

CloseBracket

]

CloseParen

)

Coalesce

??

Colon

:

Comma

,

Dec

--

Div

/

Dot

.

Eq

==

GreaterThan

>

GreaterThanOrEq

>=

Inc

++

LeftSh

<<

LessThan

<

LessThanOrEq

<=

Mod

%

Mul

*

Neg

~

Not

!

NotEq

!=

OpenBlock

{

OpenBracket

[

OpenParen

(

Optional

?.

Or

|

Exp

**

Question

?

RightSh

>>

Semicolon

;

Spread

...

StrictEq

===

StrictNotEq

!==

Sub

-

URightSh

>>>

Xor

^

Implementations

impl Punctuator

pub const fn as_assign_op(self) -> Option<AssignOp>

Attempts to convert a punctuator (+, =…) to an Assign Operator

If there is no match, None will be returned.

pub const fn as_binary_op(self) -> Option<BinaryOp>

Attempts to convert a punctuator (+, =…) to a Binary Operator

If there is no match, None will be returned.

pub const fn as_str(self) -> &'static str

Retrieves the punctuator as a static string.