Module boa_ast::expression::literal::object

Object Expression.

Structs

ObjectLiteral

Objects in ECMAScript may be defined as an unordered collection of related data, of primitive or reference types, in the form of “key: value” pairs.