Module boa_ast::expression::operator

Operator expression nodes.

An operator expression is an expression that takes an operator (such as +, typeof, +=) and one or more expressions and returns an expression as a result. An operator expression can be any of the following:

A Unary expression.

A Update expression.

A Assign expression.

A Binary expression.

A Conditional expression.

Re-exports

pub use self::assign::Assign;

pub use self::binary::Binary;

pub use self::binary::BinaryInPrivate;

pub use self::unary::Unary;

pub use self::update::Update;

Modules

assign

Assignment expression nodes, as defined by the spec.

binary

Binary expression nodes.

conditional 🔒

unary

Unary expression nodes.

update

Update expression nodes.

Structs

Conditional

The conditional (ternary) operation is the only ECMAScript operation that takes three operands.