Enum boa_ast::expression::operator::assign::AssignTarget

pub enum AssignTarget {
    Identifier(Identifier),
    Access(PropertyAccess),
    Pattern(Pattern),
}

The valid left-hand-side expressions of an assignment operator. Also called LeftHandSideExpression in the spec.

Variants

Identifier(Identifier)

A simple identifier, such as a.

Access(PropertyAccess)

A property access, such as a.prop.

Pattern(Pattern)

A pattern assignment, such as {a, b, ...c}.

Implementations

impl AssignTarget

pub fn from_expression(expression: &Expression, strict: bool) -> Option<Self>

Converts the left-hand-side Expression of an assignment expression into an AssignTarget. Returns None if the given Expression is an invalid left-hand-side for a assignment expression.

pub fn from_expression_simple( expression: &Expression, strict: bool ) -> Option<Self>

Converts the left-hand-side Expression of an assignment expression into an AssignTarget. Returns None if the given Expression is an invalid left-hand-side for a assignment expression.

The AssignmentTargetType of the expression must be simple.