Module boa_ast::expression::operator::binary

Binary expression nodes.

A Binary expression comprises any operation between two expressions (excluding assignments), such as:

Logic operations (||, &&).

Relational math (==, <).

Bit manipulation (^, |).

Arithmetic (+, %).

The comma operator (,)

Modules

op 🔒

This module implements various structure for logic handling.

Structs

Binary

Binary operations require two operands, one before the operator and one after the operator.

Enums

ArithmeticOp

Arithmetic operators take numerical values (either literals or variables) as their operands and return a single numerical value.

BinaryOp

This represents a binary operation between two values.

BitwiseOp

A bitwise operator is an operator used to perform bitwise operations on bit patterns or binary numerals that involve the manipulation of individual bits.

LogicalOp

Logical operators are typically used with Boolean (logical) values; when they are, they return a Boolean value.

RelationalOp

A relational operator compares its operands and returns a logical value based on whether the relation is true.