Module boa_ast::expression::operator::unary

Unary expression nodes.

A unary expression comprises any operation applied to a single expression. Some examples include:

The delete operator.

The bitwise NOT operator (~).

The full list of valid unary operators is defined in UnaryOp.

Modules

op 🔒

Structs

Unary

A unary expression is an operation with only one operand.

Enums

UnaryOp

A unary operator is one that takes a single operand/argument and performs an operation.