Module boa_ast::expression::operator::update

Update expression nodes.

A update expression increments or decrements it’s operand and returns a value

Increment and decrement operations (++, --).

The full list of valid update operators is defined in UpdateOp.

Modules

op 🔒

Structs

Update

A update expression is an operation with only one operand.

Enums

UpdateOp

A update operator is one that takes a single operand/argument and performs an operation.

UpdateTarget

A update expression can only be performed on identifier expressions or property access expressions.

Constants

RECURSIVE_COUNT_Update 🔒

RECURSIVE_COUNT_UpdateTarget 🔒

Statics

__KEY 🔒

__KEY 🔒

Functions

__getit 🔒 ⚠

__getit 🔒 ⚠

__init 🔒

__init 🔒