Function boa_ast::block_to_string

pub(crate) fn block_to_string(
    body: &StatementList,
    interner: &Interner,
    indentation: usize
) -> String

Displays the body of a block or statement list.

This includes the curly braces at the start and end. This will not indent the first brace, but will indent the last brace.