Struct boa_ast::function::parameters::FormalParameterList

pub struct FormalParameterList {
    parameters: Box<[FormalParameter]>,
    flags: FormalParameterListFlags,
    length: u32,
}

A list of FormalParameters that describes the parameters of a function, as defined by the spec.

Fields
parameters: Box<[FormalParameter]>flags: FormalParameterListFlagslength: u32
Implementations

impl FormalParameterList

pub fn new() -> Self

Creates a new empty formal parameter list.

pub fn from_parameters(parameters: Vec<FormalParameter>) -> Self

Creates a FormalParameterList from a list of FormalParameters.

pub const fn length(&self) -> u32

Returns the length of the parameter list. Note that this is not equal to the length of the parameters slice.

pub const fn flags(&self) -> FormalParameterListFlags

Returns the parameter list flags.

pub const fn is_simple(&self) -> bool

Indicates if the parameter list is simple.

pub const fn has_duplicates(&self) -> bool

Indicates if the parameter list has duplicate parameters.

pub const fn has_rest_parameter(&self) -> bool

Indicates if the parameter list has a rest parameter.

pub const fn has_expressions(&self) -> bool

Indicates if the parameter list has expressions in it’s parameters.

pub const fn has_arguments(&self) -> bool

Indicates if the parameter list has parameters named ‘arguments’.