Type Alias boa_ast::function::FunctionBody

pub type FunctionBody = Script;

A Function body.

Since Script and FunctionBody has the same semantics, this is currently only an alias of the former.

More information:

ECMAScript reference

Aliased Type

struct FunctionBody {
    statements: StatementList,
}

Fields
statements: StatementList
Implementations

impl Script

pub const fn new(statements: StatementList) -> Self

Creates a new ScriptNode.

pub const fn statements(&self) -> &StatementList

Gets the list of statements of this ScriptNode.

pub fn statements_mut(&mut self) -> &mut StatementList

Gets a mutable reference to the list of statements of this ScriptNode.

pub const fn strict(&self) -> bool

Gets the strict mode.