Module boa_ast::keyword

The Keyword AST node, which represents reserved words of the ECMAScript language.

The specification defines keywords as tokens that match an IdentifierName, but also have special meaning in ECMAScript. In ECMAScript, you cannot use these reserved words as variables, labels, or function names.

The MDN documentation contains a more extensive explanation about keywords.

Structs

KeywordError

The error type which is returned from parsing a str into a Keyword.

Enums

Keyword

List of keywords recognized by the JavaScript grammar.