Function boa_ast::operations::check_labels

pub fn check_labels<N>(node: &N) -> Result<(), CheckLabelsError>
where
    N: VisitWith,

This function checks multiple syntax errors conditions for labels, break and continue.

The following syntax errors are checked:

ContainsDuplicateLabels

ContainsUndefinedBreakTarget

ContainsUndefinedContinueTarget

Early errors for BreakStatement

Early errors for ContinueStatement

Errors

This function returns an error for the first syntax error that is found.