Enum boa_ast::pattern::ArrayPatternElement

pub enum ArrayPatternElement {
    Elision,
    SingleName {
        ident: Identifier,
        default_init: Option<Expression>,
    },
    PropertyAccess {
        access: PropertyAccess,
    },
    Pattern {
        pattern: Pattern,
        default_init: Option<Expression>,
    },
    SingleNameRest {
        ident: Identifier,
    },
    PropertyAccessRest {
        access: PropertyAccess,
    },
    PatternRest {
        pattern: Pattern,
    },
}

The different types of bindings that an array binding pattern may contain.

Corresponds to the BindingElement and the AssignmentElement nodes.

Variants

Elision

Elision represents the elision of an item in the array binding pattern.

An Elision may occur at multiple points in the pattern and may be multiple elisions. This variant strictly represents one elision. If there are multiple, this should be used multiple times.

More information:

ECMAScript reference: 13.2.4 Array Initializer - Elision

SingleName

Fields

ident: Identifier

The variable name where the index element will be stored.

default_init: Option<Expression>

An optional default value for the variable, in case the index element doesn’t exist.

SingleName represents a SingleName with an identifier and an optional default initializer.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - SingleNameBinding

PropertyAccess

Fields

access: PropertyAccess

The property access where the index element will be stored.

PropertyAccess represents a binding with a property accessor.

Note: According to the spec this is not part of an ArrayBindingPattern. This is only used when a array literal is used as the left-hand-side of an assignment expression.

More information:

ECMAScript reference

Pattern

Fields

pattern: Pattern

The pattern where the index element will be stored.

default_init: Option<Expression>

An optional default value for the pattern, in case the index element doesn’t exist.

Pattern represents a Pattern in an Element of an array pattern.

The pattern and the optional default initializer are both stored in the DeclarationPattern.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - BindingElement

SingleNameRest

Fields

ident: Identifier

The variable where the unassigned index elements will be stored.

SingleNameRest represents a BindingIdentifier in a BindingRestElement of an array pattern.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - BindingRestElement

PropertyAccessRest

Fields

access: PropertyAccess

The property access where the unassigned index elements will be stored.

PropertyAccess represents a rest (spread operator) with a property accessor.

Note: According to the spec this is not part of an ArrayBindingPattern. This is only used when a array literal is used as the left-hand-side of an assignment expression.

More information:

ECMAScript reference

PatternRest

Fields

pattern: Pattern

The pattern where the unassigned index elements will be stored.

PatternRest represents a Pattern in a RestElement of an array pattern.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - BindingRestElement