Struct boa_ast::pattern::ObjectPattern

pub struct ObjectPattern(Box<[ObjectPatternElement]>);

An object binding or assignment pattern.

Corresponds to the ObjectBindingPattern and the ObjectAssignmentPattern Parse Nodes.

For more information on what is a valid binding in an object pattern, see ObjectPatternElement.

Tuple Fields
0: Box<[ObjectPatternElement]>
Implementations

impl ObjectPattern

pub fn new(bindings: Box<[ObjectPatternElement]>) -> Self

Creates a new object binding pattern.

pub const fn bindings(&self) -> &[ObjectPatternElement]

Gets the bindings for the object binding pattern.

pub const fn has_rest(&self) -> bool

Returns true if the object binding pattern has a rest element.