Module boa_ast::statement::iteration::for_loop

Structs

ForLoop

The for statement creates a loop that consists of three optional expressions.

InnerForLoop 🔒

Inner structure to avoid multiple indirections in the heap.

Enums

ForLoopInitializer

A ForLoop initializer, as defined by the spec.

Constants

RECURSIVE_COUNT_ForLoop 🔒

RECURSIVE_COUNT_ForLoopInitializer 🔒

RECURSIVE_COUNT_InnerForLoop 🔒

Statics

__KEY 🔒

__KEY 🔒

__KEY 🔒

Functions

__getit 🔒 ⚠

__getit 🔒 ⚠

__getit 🔒 ⚠

__init 🔒

__init 🔒

__init 🔒