Struct boa_ast::statement::If

pub struct If {
    condition: Expression,
    body: Box<Statement>,
    else_node: Option<Box<Statement>>,
}

The if statement executes a statement if a specified condition is truthy. If the condition is falsy, another statement can be executed.

Multiple if...else statements can be nested to create an else if clause.

Note that there is no elseif (in one word) keyword in JavaScript.

More information:

ECMAScript reference

MDN documentation

Fields
condition: Expressionbody: Box<Statement>else_node: Option<Box<Statement>>
Implementations

impl If

pub const fn cond(&self) -> &Expression

Gets the condition of the if statement.

pub const fn body(&self) -> &Statement

Gets the body to execute if the condition is true.

pub fn else_node(&self) -> Option<&Statement>

Gets the else node, if it has one.

pub fn new(
condition: Expression,
body: Statement,
else_node: Option<Statement>
) -> Self

Creates an If AST node.