Struct boa_ast::statement::switch::Switch

pub struct Switch {
    val: Expression,
    cases: Box<[Case]>,
    default: Option<StatementList>,
}

The switch statement evaluates an expression, matching the expression’s value to a case clause, and executes statements associated with that case, as well as statements in cases that follow the matching case.

A switch statement first evaluates its expression. It then looks for the first case clause whose expression evaluates to the same value as the result of the input expression (using the strict comparison, ===) and transfers control to that clause, executing the associated statements. (If multiple cases match the provided value, the first case that matches is selected, even if the cases are not equal to each other.)

More information:

ECMAScript reference

MDN documentation

Fields
val: Expressioncases: Box<[Case]>default: Option<StatementList>
Implementations

impl Switch

pub fn new(
val: Expression,
cases: Box<[Case]>,
default: Option<StatementList>
) -> Self

Creates a Switch AST node.

pub fn val(&self) -> &Expression

Gets the value to switch.

pub fn cases(&self) -> &[Case]

Gets the list of cases for the switch statement.

pub fn default(&self) -> Option<&StatementList>

Gets the default statement list, if any.

pub(crate) fn contains_arguments(&self) -> bool

pub(crate) fn contains(&self, symbol: ContainsSymbol) -> bool