Struct boa_ast::ModuleItemList

pub struct ModuleItemList {
    items: Box<[ModuleItem]>,
}

Module item list AST node.

It contains a list of module items.

More information:

ECMAScript specification

Fields
items: Box<[ModuleItem]>
Implementations

impl ModuleItemList

pub const fn items(&self) -> &[ModuleItem]

Gets the list of module items.

pub fn exported_names(&self) -> Vec<Sym>

Abstract operation ExportedNames.

pub fn exported_bindings(&self) -> FxHashSet<Identifier>

Abstract operation ExportedBindings.

pub fn requests(&self) -> IndexSet<Sym, BuildHasherDefault<FxHasher>>

Operation ModuleRequests.

Gets the list of modules that need to be fetched by the module resolver to link this module.

pub fn import_entries(&self) -> Vec<ImportEntry>

Operation ImportEntries.

Gets the list of import entries of this module.

pub fn export_entries(&self) -> Vec<ExportEntry>

Operation ExportEntries.

Gets the list of export entries of this module.