Trait boa_ast::visitor::Visitor

pub trait Visitor<'ast>: Sized {
    type BreakTy;

    // Provided methods
    fn visit_script(&mut self, node: &'ast Script) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_module(&mut self, node: &'ast Module) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_statement_list(
        &mut self,
        node: &'ast StatementList
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_statement_list_item(
        &mut self,
        node: &'ast StatementListItem
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_statement(
        &mut self,
        node: &'ast Statement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_declaration(
        &mut self,
        node: &'ast Declaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_function(
        &mut self,
        node: &'ast Function
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_generator(
        &mut self,
        node: &'ast Generator
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_async_function(
        &mut self,
        node: &'ast AsyncFunction
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_async_generator(
        &mut self,
        node: &'ast AsyncGenerator
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_class(&mut self, node: &'ast Class) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_lexical_declaration(
        &mut self,
        node: &'ast LexicalDeclaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_block(&mut self, node: &'ast Block) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_var_declaration(
        &mut self,
        node: &'ast VarDeclaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_expression(
        &mut self,
        node: &'ast Expression
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_if(&mut self, node: &'ast If) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_do_while_loop(
        &mut self,
        node: &'ast DoWhileLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_while_loop(
        &mut self,
        node: &'ast WhileLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_for_loop(
        &mut self,
        node: &'ast ForLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_for_in_loop(
        &mut self,
        node: &'ast ForInLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_for_of_loop(
        &mut self,
        node: &'ast ForOfLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_switch(&mut self, node: &'ast Switch) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_continue(
        &mut self,
        node: &'ast Continue
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_break(&mut self, node: &'ast Break) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_return(&mut self, node: &'ast Return) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_labelled(
        &mut self,
        node: &'ast Labelled
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_throw(&mut self, node: &'ast Throw) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_try(&mut self, node: &'ast Try) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_with(&mut self, node: &'ast With) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_identifier(
        &mut self,
        node: &'ast Identifier
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_formal_parameter_list(
        &mut self,
        node: &'ast FormalParameterList
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_class_element(
        &mut self,
        node: &'ast ClassElement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_private_name(
        &mut self,
        node: &'ast PrivateName
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_variable_list(
        &mut self,
        node: &'ast VariableList
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_variable(
        &mut self,
        node: &'ast Variable
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_binding(
        &mut self,
        node: &'ast Binding
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_pattern(
        &mut self,
        node: &'ast Pattern
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_literal(
        &mut self,
        node: &'ast Literal
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_reg_exp_literal(
        &mut self,
        node: &'ast RegExpLiteral
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_array_literal(
        &mut self,
        node: &'ast ArrayLiteral
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_object_literal(
        &mut self,
        node: &'ast ObjectLiteral
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_spread(&mut self, node: &'ast Spread) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_arrow_function(
        &mut self,
        node: &'ast ArrowFunction
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_async_arrow_function(
        &mut self,
        node: &'ast AsyncArrowFunction
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_template_literal(
        &mut self,
        node: &'ast TemplateLiteral
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_property_access(
        &mut self,
        node: &'ast PropertyAccess
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_new(&mut self, node: &'ast New) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_call(&mut self, node: &'ast Call) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_super_call(
        &mut self,
        node: &'ast SuperCall
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_import_call(
        &mut self,
        node: &'ast ImportCall
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_optional(
        &mut self,
        node: &'ast Optional
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_tagged_template(
        &mut self,
        node: &'ast TaggedTemplate
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_assign(&mut self, node: &'ast Assign) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_unary(&mut self, node: &'ast Unary) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_update(&mut self, node: &'ast Update) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_binary(&mut self, node: &'ast Binary) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_binary_in_private(
        &mut self,
        node: &'ast BinaryInPrivate
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_conditional(
        &mut self,
        node: &'ast Conditional
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_await(&mut self, node: &'ast Await) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_yield(&mut self, node: &'ast Yield) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_parenthesized(
        &mut self,
        node: &'ast Parenthesized
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_for_loop_initializer(
        &mut self,
        node: &'ast ForLoopInitializer
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_iterable_loop_initializer(
        &mut self,
        node: &'ast IterableLoopInitializer
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_case(&mut self, node: &'ast Case) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_sym(&mut self, node: &'ast Sym) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_labelled_item(
        &mut self,
        node: &'ast LabelledItem
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_catch(&mut self, node: &'ast Catch) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_finally(
        &mut self,
        node: &'ast Finally
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_formal_parameter(
        &mut self,
        node: &'ast FormalParameter
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_property_name(
        &mut self,
        node: &'ast PropertyName
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_method_definition(
        &mut self,
        node: &'ast MethodDefinition
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_object_pattern(
        &mut self,
        node: &'ast ObjectPattern
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_array_pattern(
        &mut self,
        node: &'ast ArrayPattern
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_property_definition(
        &mut self,
        node: &'ast PropertyDefinition
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_template_element(
        &mut self,
        node: &'ast TemplateElement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_simple_property_access(
        &mut self,
        node: &'ast SimplePropertyAccess
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_private_property_access(
        &mut self,
        node: &'ast PrivatePropertyAccess
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_super_property_access(
        &mut self,
        node: &'ast SuperPropertyAccess
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_optional_operation(
        &mut self,
        node: &'ast OptionalOperation
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_assign_target(
        &mut self,
        node: &'ast AssignTarget
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_object_pattern_element(
        &mut self,
        node: &'ast ObjectPatternElement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_array_pattern_element(
        &mut self,
        node: &'ast ArrayPatternElement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_property_access_field(
        &mut self,
        node: &'ast PropertyAccessField
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_optional_operation_kind(
        &mut self,
        node: &'ast OptionalOperationKind
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_module_item_list(
        &mut self,
        node: &'ast ModuleItemList
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_module_item(
        &mut self,
        node: &'ast ModuleItem
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_module_specifier(
        &mut self,
        node: &'ast ModuleSpecifier
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_import_kind(
        &mut self,
        node: &'ast ImportKind
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_import_declaration(
        &mut self,
        node: &'ast ImportDeclaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_import_specifier(
        &mut self,
        node: &'ast ImportSpecifier
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_re_export_kind(
        &mut self,
        node: &'ast ReExportKind
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_export_declaration(
        &mut self,
        node: &'ast ExportDeclaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_export_specifier(
        &mut self,
        node: &'ast ExportSpecifier
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit<N: Into<NodeRef<'ast>>>(
        &mut self,
        node: N
    ) -> ControlFlow<Self::BreakTy> { ... }

}

Represents an AST visitor.

This implementation is based largely on chalk’s visitor pattern.

Required Associated Types

type BreakTy

Type which will be propagated from the visitor if completing early.

Provided Methods

fn visit_script(&mut self, node: &'ast Script) -> ControlFlow<Self::BreakTy>

Visits a Script with this visitor

fn visit_module(&mut self, node: &'ast Module) -> ControlFlow<Self::BreakTy>

Visits a Module with this visitor

fn visit_statement_list( &mut self, node: &'ast StatementList ) -> ControlFlow<Self::BreakTy>

Visits a StatementList with this visitor

fn visit_statement_list_item( &mut self, node: &'ast StatementListItem ) -> ControlFlow<Self::BreakTy>

Visits a StatementListItem with this visitor

fn visit_statement( &mut self, node: &'ast Statement ) -> ControlFlow<Self::BreakTy>

Visits a Statement with this visitor

fn visit_declaration( &mut self, node: &'ast Declaration ) -> ControlFlow<Self::BreakTy>

Visits a Declaration with this visitor

fn visit_function(&mut self, node: &'ast Function) -> ControlFlow<Self::BreakTy>

Visits a Function with this visitor

fn visit_generator( &mut self, node: &'ast Generator ) -> ControlFlow<Self::BreakTy>

Visits a Generator with this visitor

fn visit_async_function( &mut self, node: &'ast AsyncFunction ) -> ControlFlow<Self::BreakTy>

Visits a AsyncFunction with this visitor

fn visit_async_generator( &mut self, node: &'ast AsyncGenerator ) -> ControlFlow<Self::BreakTy>

Visits a AsyncGenerator with this visitor

fn visit_class(&mut self, node: &'ast Class) -> ControlFlow<Self::BreakTy>

Visits a Class with this visitor

fn visit_lexical_declaration( &mut self, node: &'ast LexicalDeclaration ) -> ControlFlow<Self::BreakTy>

Visits a LexicalDeclaration with this visitor

fn visit_block(&mut self, node: &'ast Block) -> ControlFlow<Self::BreakTy>

Visits a Block with this visitor

fn visit_var_declaration( &mut self, node: &'ast VarDeclaration ) -> ControlFlow<Self::BreakTy>

Visits a VarDeclaration with this visitor

fn visit_expression( &mut self, node: &'ast Expression ) -> ControlFlow<Self::BreakTy>

Visits a Expression with this visitor

fn visit_if(&mut self, node: &'ast If) -> ControlFlow<Self::BreakTy>

Visits a If with this visitor

fn visit_do_while_loop( &mut self, node: &'ast DoWhileLoop ) -> ControlFlow<Self::BreakTy>

Visits a DoWhileLoop with this visitor

fn visit_while_loop( &mut self, node: &'ast WhileLoop ) -> ControlFlow<Self::BreakTy>

Visits a WhileLoop with this visitor

fn visit_for_loop(&mut self, node: &'ast ForLoop) -> ControlFlow<Self::BreakTy>

Visits a ForLoop with this visitor

fn visit_for_in_loop( &mut self, node: &'ast ForInLoop ) -> ControlFlow<Self::BreakTy>

Visits a ForInLoop with this visitor

fn visit_for_of_loop( &mut self, node: &'ast ForOfLoop ) -> ControlFlow<Self::BreakTy>

Visits a ForOfLoop with this visitor

fn visit_switch(&mut self, node: &'ast Switch) -> ControlFlow<Self::BreakTy>

Visits a Switch with this visitor

fn visit_continue(&mut self, node: &'ast Continue) -> ControlFlow<Self::BreakTy>

Visits a Continue with this visitor

fn visit_break(&mut self, node: &'ast Break) -> ControlFlow<Self::BreakTy>

Visits a Break with this visitor

fn visit_return(&mut self, node: &'ast Return) -> ControlFlow<Self::BreakTy>

Visits a Return with this visitor

fn visit_labelled(&mut self, node: &'ast Labelled) -> ControlFlow<Self::BreakTy>

Visits a Labelled with this visitor

fn visit_throw(&mut self, node: &'ast Throw) -> ControlFlow<Self::BreakTy>

Visits a Throw with this visitor

fn visit_try(&mut self, node: &'ast Try) -> ControlFlow<Self::BreakTy>

Visits a Try with this visitor

fn visit_with(&mut self, node: &'ast With) -> ControlFlow<Self::BreakTy>

Visits a With with this visitor

fn visit_identifier( &mut self, node: &'ast Identifier ) -> ControlFlow<Self::BreakTy>

Visits a Identifier with this visitor

fn visit_formal_parameter_list( &mut self, node: &'ast FormalParameterList ) -> ControlFlow<Self::BreakTy>

Visits a FormalParameterList with this visitor

fn visit_class_element( &mut self, node: &'ast ClassElement ) -> ControlFlow<Self::BreakTy>

Visits a ClassElement with this visitor

fn visit_private_name( &mut self, node: &'ast PrivateName ) -> ControlFlow<Self::BreakTy>

Visits a PrivateName with this visitor

fn visit_variable_list( &mut self, node: &'ast VariableList ) -> ControlFlow<Self::BreakTy>

Visits a VariableList with this visitor

fn visit_variable(&mut self, node: &'ast Variable) -> ControlFlow<Self::BreakTy>

Visits a Variable with this visitor

fn visit_binding(&mut self, node: &'ast Binding) -> ControlFlow<Self::BreakTy>

Visits a Binding with this visitor

fn visit_pattern(&mut self, node: &'ast Pattern) -> ControlFlow<Self::BreakTy>

Visits a Pattern with this visitor

fn visit_literal(&mut self, node: &'ast Literal) -> ControlFlow<Self::BreakTy>

Visits a Literal with this visitor

fn visit_reg_exp_literal( &mut self, node: &'ast RegExpLiteral ) -> ControlFlow<Self::BreakTy>

Visits a RegExpLiteral with this visitor

fn visit_array_literal( &mut self, node: &'ast ArrayLiteral ) -> ControlFlow<Self::BreakTy>

Visits a ArrayLiteral with this visitor

fn visit_object_literal( &mut self, node: &'ast ObjectLiteral ) -> ControlFlow<Self::BreakTy>

Visits a ObjectLiteral with this visitor

fn visit_spread(&mut self, node: &'ast Spread) -> ControlFlow<Self::BreakTy>

Visits a Spread with this visitor

fn visit_arrow_function( &mut self, node: &'ast ArrowFunction ) -> ControlFlow<Self::BreakTy>

Visits a ArrowFunction with this visitor

fn visit_async_arrow_function( &mut self, node: &'ast AsyncArrowFunction ) -> ControlFlow<Self::BreakTy>

Visits a AsyncArrowFunction with this visitor

fn visit_template_literal( &mut self, node: &'ast TemplateLiteral ) -> ControlFlow<Self::BreakTy>

Visits a TemplateLiteral with this visitor

fn visit_property_access( &mut self, node: &'ast PropertyAccess ) -> ControlFlow<Self::BreakTy>

Visits a PropertyAccess with this visitor

fn visit_new(&mut self, node: &'ast New) -> ControlFlow<Self::BreakTy>

Visits a New with this visitor

fn visit_call(&mut self, node: &'ast Call) -> ControlFlow<Self::BreakTy>

Visits a Call with this visitor

fn visit_super_call( &mut self, node: &'ast SuperCall ) -> ControlFlow<Self::BreakTy>

Visits a SuperCall with this visitor

fn visit_import_call( &mut self, node: &'ast ImportCall ) -> ControlFlow<Self::BreakTy>

Visits a ImportCall with this visitor

fn visit_optional(&mut self, node: &'ast Optional) -> ControlFlow<Self::BreakTy>

Visits a Optional with this visitor

fn visit_tagged_template( &mut self, node: &'ast TaggedTemplate ) -> ControlFlow<Self::BreakTy>

Visits a TaggedTemplate with this visitor

fn visit_assign(&mut self, node: &'ast Assign) -> ControlFlow<Self::BreakTy>

Visits a Assign with this visitor

fn visit_unary(&mut self, node: &'ast Unary) -> ControlFlow<Self::BreakTy>

Visits a Unary with this visitor

fn visit_update(&mut self, node: &'ast Update) -> ControlFlow<Self::BreakTy>

Visits a Update with this visitor

fn visit_binary(&mut self, node: &'ast Binary) -> ControlFlow<Self::BreakTy>

Visits a Binary with this visitor

fn visit_binary_in_private( &mut self, node: &'ast BinaryInPrivate ) -> ControlFlow<Self::BreakTy>

Visits a BinaryInPrivate with this visitor

fn visit_conditional( &mut self, node: &'ast Conditional ) -> ControlFlow<Self::BreakTy>

Visits a Conditional with this visitor

fn visit_await(&mut self, node: &'ast Await) -> ControlFlow<Self::BreakTy>

Visits a Await with this visitor

fn visit_yield(&mut self, node: &'ast Yield) -> ControlFlow<Self::BreakTy>

Visits a Yield with this visitor

fn visit_parenthesized( &mut self, node: &'ast Parenthesized ) -> ControlFlow<Self::BreakTy>

Visits a Parenthesized with this visitor

fn visit_for_loop_initializer( &mut self, node: &'ast ForLoopInitializer ) -> ControlFlow<Self::BreakTy>

Visits a ForLoopInitializer with this visitor

fn visit_iterable_loop_initializer( &mut self, node: &'ast IterableLoopInitializer ) -> ControlFlow<Self::BreakTy>

Visits a IterableLoopInitializer with this visitor

fn visit_case(&mut self, node: &'ast Case) -> ControlFlow<Self::BreakTy>

Visits a Case with this visitor

fn visit_sym(&mut self, node: &'ast Sym) -> ControlFlow<Self::BreakTy>

Visits a Sym with this visitor

fn visit_labelled_item( &mut self, node: &'ast LabelledItem ) -> ControlFlow<Self::BreakTy>

Visits a LabelledItem with this visitor

fn visit_catch(&mut self, node: &'ast Catch) -> ControlFlow<Self::BreakTy>

Visits a Catch with this visitor

fn visit_finally(&mut self, node: &'ast Finally) -> ControlFlow<Self::BreakTy>

Visits a Finally with this visitor

fn visit_formal_parameter( &mut self, node: &'ast FormalParameter ) -> ControlFlow<Self::BreakTy>

Visits a FormalParameter with this visitor

fn visit_property_name( &mut self, node: &'ast PropertyName ) -> ControlFlow<Self::BreakTy>

Visits a PropertyName with this visitor

fn visit_method_definition( &mut self, node: &'ast MethodDefinition ) -> ControlFlow<Self::BreakTy>

Visits a MethodDefinition with this visitor

fn visit_object_pattern( &mut self, node: &'ast ObjectPattern ) -> ControlFlow<Self::BreakTy>

Visits a ObjectPattern with this visitor

fn visit_array_pattern( &mut self, node: &'ast ArrayPattern ) -> ControlFlow<Self::BreakTy>

Visits a ArrayPattern with this visitor

fn visit_property_definition( &mut self, node: &'ast PropertyDefinition ) -> ControlFlow<Self::BreakTy>

Visits a PropertyDefinition with this visitor

fn visit_template_element( &mut self, node: &'ast TemplateElement ) -> ControlFlow<Self::BreakTy>

Visits a TemplateElement with this visitor

fn visit_simple_property_access( &mut self, node: &'ast SimplePropertyAccess ) -> ControlFlow<Self::BreakTy>

Visits a SimplePropertyAccess with this visitor

fn visit_private_property_access( &mut self, node: &'ast PrivatePropertyAccess ) -> ControlFlow<Self::BreakTy>

Visits a PrivatePropertyAccess with this visitor

fn visit_super_property_access( &mut self, node: &'ast SuperPropertyAccess ) -> ControlFlow<Self::BreakTy>

Visits a SuperPropertyAccess with this visitor

fn visit_optional_operation( &mut self, node: &'ast OptionalOperation ) -> ControlFlow<Self::BreakTy>

Visits a OptionalOperation with this visitor

fn visit_assign_target( &mut self, node: &'ast AssignTarget ) -> ControlFlow<Self::BreakTy>

Visits a AssignTarget with this visitor

fn visit_object_pattern_element( &mut self, node: &'ast ObjectPatternElement ) -> ControlFlow<Self::BreakTy>

Visits a ObjectPatternElement with this visitor

fn visit_array_pattern_element( &mut self, node: &'ast ArrayPatternElement ) -> ControlFlow<Self::BreakTy>

Visits a ArrayPatternElement with this visitor

fn visit_property_access_field( &mut self, node: &'ast PropertyAccessField ) -> ControlFlow<Self::BreakTy>

Visits a PropertyAccessField with this visitor

fn visit_optional_operation_kind( &mut self, node: &'ast OptionalOperationKind ) -> ControlFlow<Self::BreakTy>

Visits a OptionalOperationKind with this visitor

fn visit_module_item_list( &mut self, node: &'ast ModuleItemList ) -> ControlFlow<Self::BreakTy>

Visits a ModuleItemList with this visitor

fn visit_module_item( &mut self, node: &'ast ModuleItem ) -> ControlFlow<Self::BreakTy>

Visits a ModuleItem with this visitor

fn visit_module_specifier( &mut self, node: &'ast ModuleSpecifier ) -> ControlFlow<Self::BreakTy>

Visits a ModuleSpecifier with this visitor

fn visit_import_kind( &mut self, node: &'ast ImportKind ) -> ControlFlow<Self::BreakTy>

Visits a ImportKind with this visitor

fn visit_import_declaration( &mut self, node: &'ast ImportDeclaration ) -> ControlFlow<Self::BreakTy>

Visits a ImportDeclaration with this visitor

fn visit_import_specifier( &mut self, node: &'ast ImportSpecifier ) -> ControlFlow<Self::BreakTy>

Visits a ImportSpecifier with this visitor

fn visit_re_export_kind( &mut self, node: &'ast ReExportKind ) -> ControlFlow<Self::BreakTy>

Visits a ReExportKind with this visitor

fn visit_export_declaration( &mut self, node: &'ast ExportDeclaration ) -> ControlFlow<Self::BreakTy>

Visits a ExportDeclaration with this visitor

fn visit_export_specifier( &mut self, node: &'ast ExportSpecifier ) -> ControlFlow<Self::BreakTy>

Visits a ExportSpecifier with this visitor

fn visit<N: Into<NodeRef<'ast>>>( &mut self, node: N ) -> ControlFlow<Self::BreakTy>

Generic entry point for a node that is visitable by a Visitor.

This is usually used for generic functions that need to visit an unnamed AST node.