Module boa_ast::declaration

The Declaration Parse Node, as defined by the spec.

ECMAScript declarations include:

Lexical declarations (let, const).

Function declarations (function, async function).

Class declarations.

See Difference between statements and declarations for an explanation on why Declarations and Statements are distinct nodes.

Modules

export 🔒

Export declaration AST nodes.

import 🔒

Import declaration AST nodes.

variable 🔒

Variable related declarations.

Structs

ExportSpecifier

Export specifier

ImportDeclaration

An import declaration AST node.

ImportSpecifier

Import specifier

ModuleSpecifier

Module specifier.

TryFromVariableListError

The error returned by the VariableList::try_from function.

VarDeclaration

A var statement, also called VariableStatement in the spec.

Variable

Variable represents a variable declaration of some kind.

VariableList

List of variables in a variable declaration.

Enums

Binding

Binding represents either an individual binding or a binding pattern.

Declaration

The Declaration Parse Node.

ExportDeclaration

An export declaration AST node.

ImportKind

The kind of import in an ImportDeclaration.

LexicalDeclaration

A lexical declaration defines variables that are scoped to the lexical environment of the variable declaration.

ReExportKind

The kind of re-export in an ExportDeclaration.