Module boa_ast::declaration

The Declaration Parse Node, as defined by the spec.

Javascript declarations include:

Lexical declarations (let, const).

Function declarations (function, async function).

Class declarations.

See Difference between statements and declarations for an explanation on why Declarations and Statements are distinct nodes.

Modules

variable 🔒

Variable related declarations.

Structs

TryFromVariableListError

The error returned by the VariableList::try_from function.

VarDeclaration

A var statement, also called VariableStatement in the spec.

Variable

Variable represents a variable declaration of some kind.

VariableList

List of variables in a variable declaration.

Enums

Binding

Binding represents either an individual binding or a binding pattern.

Declaration

The Declaration Parse Node.

LexicalDeclaration

A lexical declaration defines variables that are scoped to the lexical environment of the variable declaration.