Module boa_ast::expression

The Expression Parse Node, as defined by the spec.

Javascript expressions include:

Primary expressions (this, function expressions, literals).

Left hand side expressions (accessors, new operator, super).

operator expressions.

Modules

access

Property access expressions, as defined by the spec.

await 🔒

Await expression Expression.

call 🔒

identifier 🔒

Local identifier Expression.

literal

This module contains all literal expressions, which represents the primitive values in JavaScript.

new 🔒

operator

Operator expression nodes.

optional 🔒

spread 🔒

tagged_template 🔒

yield 🔒

Structs

Await

An await expression is used within an async function to pause execution and wait for a promise to resolve.

Call

Calling the function actually performs the specified actions with the indicated parameters.

Identifier

An identifier is a sequence of characters in the code that identifies a variable, function, or property.

New

The new operator lets developers create an instance of a user-defined object type or of one of the built-in object types that has a constructor function.

Optional

An optional chain expression, as defined by the spec.

OptionalOperation

Operation within an Optional chain.

Spread

The spread operator allows an iterable such as an array expression or string to be expanded.

SuperCall

The super keyword is used to access and call functions on an object’s parent.

TaggedTemplate

A TaggedTemplate expression, as defined by the spec.

Yield

The yield keyword is used to pause and resume a generator function

Enums

Expression

The Expression Parse Node.

OptionalOperationKind

List of valid operations in an Optional chain.

Constants

RESERVED_IDENTIFIERS_STRICT

List of reserved keywords exclusive to strict mode.