Enum boa_ast::expression::Expression

pub enum Expression {

    This,
    Identifier(Identifier),
    Literal(Literal),
    ArrayLiteral(ArrayLiteral),
    ObjectLiteral(ObjectLiteral),
    Spread(Spread),
    Function(Function),
    ArrowFunction(ArrowFunction),
    AsyncArrowFunction(AsyncArrowFunction),
    Generator(Generator),
    AsyncFunction(AsyncFunction),
    AsyncGenerator(AsyncGenerator),
    Class(Box<Class>),
    TemplateLiteral(TemplateLiteral),
    PropertyAccess(PropertyAccess),
    New(New),
    Call(Call),
    SuperCall(SuperCall),
    Optional(Optional),
    TaggedTemplate(TaggedTemplate),
    NewTarget,
    Assign(Assign),
    Unary(Unary),
    Binary(Binary),
    Conditional(Conditional),
    Await(Await),
    Yield(Yield),
    // some variants omitted

}

The Expression Parse Node.

See the module level documentation for more information.

Variants

This

The JavaScript this keyword refers to the object it belongs to.

A property of an execution context (global, function or eval) that, in non–strict mode, is always a reference to an object and in strict mode can be any value.

More information:

ECMAScript reference

MDN documentation

Identifier(Identifier)

See Identifier.

Literal(Literal)

See Literal.

ArrayLiteral(ArrayLiteral)

See ArrayLiteral.

ObjectLiteral(ObjectLiteral)

See ObjectLiteral.

Spread(Spread)

See Spread,

Function(Function)

See Function.

ArrowFunction(ArrowFunction)

See ArrowFunction.

AsyncArrowFunction(AsyncArrowFunction)

See AsyncArrowFunction.

Generator(Generator)

See Generator.

AsyncFunction(AsyncFunction)

See AsyncFunction.

AsyncGenerator(AsyncGenerator)

See AsyncGenerator.

Class(Box<Class>)

See Class.

TemplateLiteral(TemplateLiteral)

See TemplateLiteral.

PropertyAccess(PropertyAccess)

See PropertyAccess.

New(New)

See New.

Call(Call)

See Call.

SuperCall(SuperCall)

See SuperCall.

Optional(Optional)

See Optional.

TaggedTemplate(TaggedTemplate)

See TaggedTemplate.

NewTarget

The new.target pseudo-property expression.

Assign(Assign)

See Assign.

Unary(Unary)

See Unary.

Binary(Binary)

See Binary.

Conditional(Conditional)

See Conditional.

Await(Await)

See Await.

Yield(Yield)

See Yield.

Implementations

impl Expression

pub(crate) fn to_no_indent_string(
&self,
interner: &Interner,
indentation: usize
) -> String

Implements the display formatting with indentation.

This will not prefix the value with any indentation. If you want to prefix this with proper indents, use to_indented_string().