Enum boa_ast::expression::Expression

pub enum Expression {

    This,
    Identifier(Identifier),
    Literal(Literal),
    ArrayLiteral(ArrayLiteral),
    ObjectLiteral(ObjectLiteral),
    Spread(Spread),
    Function(Function),
    ArrowFunction(ArrowFunction),
    AsyncArrowFunction(AsyncArrowFunction),
    Generator(Generator),
    AsyncFunction(AsyncFunction),
    AsyncGenerator(AsyncGenerator),
    Class(Box<Class>),
    TemplateLiteral(TemplateLiteral),
    PropertyAccess(PropertyAccess),
    New(New),
    Call(Call),
    SuperCall(SuperCall),
    ImportCall(ImportCall),
    Optional(Optional),
    TaggedTemplate(TaggedTemplate),
    NewTarget,
    ImportMeta,
    Assign(Assign),
    Unary(Unary),
    Update(Update),
    Binary(Binary),
    BinaryInPrivate(BinaryInPrivate),
    Conditional(Conditional),
    Await(Await),
    Yield(Yield),
    Parenthesized(Parenthesized),
    // some variants omitted

}

The Expression Parse Node.

See the module level documentation for more information.

Variants

This

The ECMAScript this keyword refers to the object it belongs to.

A property of an execution context (global, function or eval) that, in non–strict mode, is always a reference to an object and in strict mode can be any value.

More information:

ECMAScript reference

MDN documentation

Identifier(Identifier)

See Identifier.

Literal(Literal)

See Literal.

ArrayLiteral(ArrayLiteral)

See ArrayLiteral.

ObjectLiteral(ObjectLiteral)

See ObjectLiteral.

Spread(Spread)

See Spread,

Function(Function)

See Function.

ArrowFunction(ArrowFunction)

See ArrowFunction.

AsyncArrowFunction(AsyncArrowFunction)

See AsyncArrowFunction.

Generator(Generator)

See Generator.

AsyncFunction(AsyncFunction)

See AsyncFunction.

AsyncGenerator(AsyncGenerator)

See AsyncGenerator.

Class(Box<Class>)

See Class.

TemplateLiteral(TemplateLiteral)

See TemplateLiteral.

PropertyAccess(PropertyAccess)

See PropertyAccess.

New(New)

See New.

Call(Call)

See Call.

SuperCall(SuperCall)

See SuperCall.

ImportCall(ImportCall)

See ImportCall.

Optional(Optional)

See Optional.

TaggedTemplate(TaggedTemplate)

See TaggedTemplate.

NewTarget

The new.target pseudo-property expression.

ImportMeta

The import.meta pseudo-property expression.

Assign(Assign)

See Assign.

Unary(Unary)

See Unary.

Update(Update)

See Unary.

Binary(Binary)

See Binary.

BinaryInPrivate(BinaryInPrivate)

See BinaryInPrivate.

Conditional(Conditional)

See Conditional.

Await(Await)

See Await.

Yield(Yield)

See Yield.

Parenthesized(Parenthesized)

See Parenthesized.

Implementations

impl Expression

pub const fn is_function_definition(&self) -> bool

Returns if the expression is a function definition according to the spec.

More information:

ECMAScript reference

pub const fn is_anonymous_function_definition(&self) -> bool

Returns if the expression is a function definition without a name.

More information:

ECMAScript reference

pub const fn flatten(&self) -> &Self

Returns the expression without any outer parenthesized expressions.