Module boa_ast::expression::operator::binary::op

This module implements various structure for logic handling.

Enums

ArithmeticOp

Arithmetic operators take numerical values (either literals or variables) as their operands and return a single numerical value.

BinaryOp

This represents a binary operation between two values.

BitwiseOp

A bitwise operator is an operator used to perform bitwise operations on bit patterns or binary numerals that involve the manipulation of individual bits.

LogicalOp

Logical operators are typically used with Boolean (logical) values; when they are, they return a Boolean value.

RelationalOp

A relational operator compares its operands and returns a logical value based on whether the relation is true.

Constants

RECURSIVE_COUNT_ArithmeticOp 🔒

RECURSIVE_COUNT_BinaryOp 🔒

RECURSIVE_COUNT_BitwiseOp 🔒

RECURSIVE_COUNT_LogicalOp 🔒

RECURSIVE_COUNT_RelationalOp 🔒

Statics

__KEY 🔒

__KEY 🔒

__KEY 🔒

__KEY 🔒

__KEY 🔒

Functions

__getit 🔒 ⚠

__getit 🔒 ⚠

__getit 🔒 ⚠

__getit 🔒 ⚠

__getit 🔒 ⚠

__init 🔒

__init 🔒

__init 🔒

__init 🔒

__init 🔒