Struct boa_ast::function::AsyncArrowFunction

pub struct AsyncArrowFunction {
    name: Option<Identifier>,
    parameters: FormalParameterList,
    body: FunctionBody,
}

An async arrow function expression, as defined by the spec.

An async arrow function expression is a syntactically compact alternative to a regular function expression. Arrow function expressions are ill suited as methods, and they cannot be used as constructors. Arrow functions cannot be used as constructors and will throw an error when used with new.

Fields
name: Option<Identifier>parameters: FormalParameterListbody: FunctionBody
Implementations

impl AsyncArrowFunction

pub const fn new( name: Option<Identifier>, parameters: FormalParameterList, body: FunctionBody ) -> Self

Creates a new AsyncArrowFunction AST Expression.

pub const fn name(&self) -> Option<Identifier>

Gets the name of the function declaration.

pub fn set_name(&mut self, name: Option<Identifier>)

Sets the name of the function declaration.

pub const fn parameters(&self) -> &FormalParameterList

Gets the list of parameters of the arrow function.

pub const fn body(&self) -> &FunctionBody

Gets the body of the arrow function.