Struct boa_ast::function::FormalParameter

pub struct FormalParameter {
    variable: Variable,
    is_rest_param: bool,
}

“Formal parameter” is a fancy way of saying “function parameter”.

In the declaration of a function, the parameters must be identifiers, not any value like numbers, strings, or objects.

function foo(formalParameter1, formalParameter2) {
}

More information:

ECMAScript reference

MDN documentation

Fields
variable: Variableis_rest_param: bool
Implementations

impl FormalParameter

pub fn new<D>(variable: D, is_rest_param: bool) -> Selfwhere
D: Into<Variable>,

Creates a new formal parameter.

pub const fn variable(&self) -> &Variable

Gets the variable of the formal parameter

pub const fn init(&self) -> Option<&Expression>

Gets the initialization node of the formal parameter, if any.

pub const fn is_rest_param(&self) -> bool

Returns true if the parameter is a rest parameter.

pub const fn is_identifier(&self) -> bool

Returns true if the parameter is an identifier.