Module boa_ast::operations

Definitions of various Syntax-Directed Operations used in the spec.

Structs

AllPrivateIdentifiersValidVisitor 🔒

AnnexBFunctionDeclarationNamesVisitor 🔒

The Visitor used for annex_b_function_declarations_names.

BoundNamesVisitor 🔒

The Visitor used to obtain the bound names of a node.

LexicallyDeclaredNamesVisitor 🔒

The Visitor used to obtain the lexically declared names of a node.

LexicallyScopedDeclarationsVisitor 🔒

The Visitor used to obtain the lexically scoped declarations of a node.

ReturnsValueVisitor 🔒

The Visitor used for returns_value.

TopLevelLexicallyScopedDeclarationsVisitor 🔒

The Visitor used to obtain the top level lexically scoped declarations of a node.

TopLevelVarScopedDeclarationsVisitor 🔒

The Visitor used to obtain the top level var scoped declarations of a node.

VarDeclaredNamesVisitor 🔒

The Visitor used to obtain the var declared names of a node.

VarScopedDeclarationsVisitor 🔒

The Visitor used to obtain the var scoped declarations of a node.

Enums

CheckLabelsError

Errors that can occur when checking labels.

ContainsSymbol

Represents all the possible symbols searched for by the Contains operation.

LexicallyScopedDeclaration

The type of a lexically scoped declaration.

VarScopedDeclaration

The type of a var scoped declaration.

Traits

IdentList 🔒

A container that BoundNamesVisitor can use to push the found identifiers.

Functions

all_private_identifiers_valid

Returns true if all private identifiers in a node are valid.

annex_b_function_declarations_names

Returns a list function declaration names that are directly contained in a statement lists Block, CaseClause or DefaultClause. If the function declaration would cause an early error it is not included in the list.

bound_names

Returns a list with the bound names of an AST node, which may contain duplicates.

check_labels

This function checks multiple syntax errors conditions for labels, break and continue.

contains

Returns true if the node contains the given symbol.

contains_arguments

Returns true if the node contains an identifier reference with name arguments.

contains_invalid_object_literal

Returns true if the given node contains a CoverInitializedName.

has_direct_super

Returns true if method has a super call in its parameters or body.

lexically_declared_names

Returns a list with the lexical bindings of a node, which may contain duplicates.

lexically_declared_names_legacy

Returns a list with the lexical bindings of a node, which may contain duplicates.

lexically_scoped_declarations

Returns a list of lexically scoped declarations of the given node.

returns_value

Returns true if the given statement returns a value.

top_level_lexicals 🔒

Utility function that collects the top level lexicals of a statement list into names.

top_level_vars 🔒

Utility function that collects the top level vars of a statement list into names.

var_declared_names

Returns a set with the var bindings of a node, with no duplicates.

var_scoped_declarations

Returns a list of var scoped declarations of the given node.