Function boa_ast::operations::annex_b_function_declarations_names

pub fn annex_b_function_declarations_names<'a, N>(
    node: &'a N
) -> Vec<Identifier>where
    &'a N: Into<NodeRef<'a>>,

Returns a list function declaration names that are directly contained in a statement lists Block, CaseClause or DefaultClause. If the function declaration would cause an early error it is not included in the list.

This behavior is used in the following annexB sections:

B.3.2.1 Changes to FunctionDeclarationInstantiation

B.3.2.2 Changes to GlobalDeclarationInstantiation

B.3.2.3 Changes to EvalDeclarationInstantiation