Module boa_ast::statement

The Statement Parse Node, as defined by the spec.

ECMAScript statements are mainly composed of control flow operations, such as If, WhileLoop, and Break. However, it also contains statements such as VarDeclaration, Block or Expression which are not strictly used for control flow.

Re-exports

pub use self::iteration::Break;

pub use self::iteration::Continue;

pub use self::iteration::DoWhileLoop;

pub use self::iteration::ForInLoop;

pub use self::iteration::ForLoop;

pub use self::iteration::ForOfLoop;

pub use self::iteration::WhileLoop;

Modules

block 🔒

Block AST node.

if 🔒

If statement

iteration

Iteration nodes

labelled 🔒

return 🔒

switch 🔒

Switch node.

throw 🔒

try 🔒

Error handling statements

with 🔒

Structs

Block

A block statement (or compound statement in other languages) is used to group zero or more statements.

Case

A case clause inside a Switch statement, as defined by the spec.

Catch

Catch block.

Finally

Finally block.

If

The if statement executes a statement if a specified condition is truthy. If the condition is falsy, another statement can be executed.

Labelled

Labelled statement nodes, as defined by the spec.

Return

The return statement ends function execution and specifies a value to be returned to the function caller.

Switch

The switch statement evaluates an expression, matching the expression’s value to a case clause, and executes statements associated with that case, as well as statements in cases that follow the matching case.

Throw

The throw statement throws a user-defined exception.

Try

The try...catch statement marks a block of statements to try and specifies a response should an exception be thrown.

With

The with statement extends the scope chain for a statement.

Enums

ErrorHandler

The type of error handler in a Try statement.

LabelledItem

The set of Parse Nodes that can be preceded by a label, as defined by the spec.

Statement

The Statement Parse Node.

Constants

RECURSIVE_COUNT_Statement 🔒

Statics

__KEY 🔒

Functions

__getit 🔒 ⚠

__init 🔒