Enum boa_ast::statement::Statement

pub enum Statement {

    Block(Block),
    Var(VarDeclaration),
    Empty,
    Expression(Expression),
    If(If),
    DoWhileLoop(DoWhileLoop),
    WhileLoop(WhileLoop),
    ForLoop(ForLoop),
    ForInLoop(ForInLoop),
    ForOfLoop(ForOfLoop),
    Switch(Switch),
    Continue(Continue),
    Break(Break),
    Return(Return),
    Labelled(Labelled),
    Throw(Throw),
    Try(Try),

}

The Statement Parse Node.

See the module level documentation for more information.

Variants

Block(Block)

See Block.

Var(VarDeclaration)

See VarDeclaration

Empty

An empty statement.

Empty statements do nothing, just return undefined.

More information:

ECMAScript reference

MDN documentation

Expression(Expression)

See Expression.

If(If)

See If.

DoWhileLoop(DoWhileLoop)

See DoWhileLoop.

WhileLoop(WhileLoop)

See WhileLoop.

ForLoop(ForLoop)

See ForLoop.

ForInLoop(ForInLoop)

See ForInLoop.

ForOfLoop(ForOfLoop)

See ForOfLoop.

Switch(Switch)

SeeSwitch.

Continue(Continue)

See Continue.

Break(Break)

See Break.

Return(Return)

See Return.

Labelled(Labelled)

See Labelled.

Throw(Throw)

See Throw.

Try(Try)

See Try.

Implementations

impl Statement

pub(crate) fn to_no_indent_string(
&self,
interner: &Interner,
indentation: usize
) -> String

Implements the display formatting with indentation.

This will not prefix the value with any indentation. If you want to prefix this with proper indents, use to_indented_string().

pub fn is_labelled_function(&self) -> bool

Abstract operation IsLabelledFunction.

This recursively checks if this Statement is a labelled function, since adding several labels in a function should not change the return value of the abstract operation:

l1: l2: l3: l4: function f(){ }

This should return true for that snippet.