Module boa_ast::statement::iteration

Iteration nodes

Modules

break 🔒

continue 🔒

do_while_loop 🔒

for_in_loop 🔒

for_loop 🔒

for_of_loop 🔒

while_loop 🔒

Structs

Break

The break statement terminates the current loop, switch, or label statement and transfers program control to the statement following the terminated statement.

Continue

The continue statement terminates execution of the statements in the current iteration of the current or labeled loop, and continues execution of the loop with the next iteration.

DoWhileLoop

The do...while statement creates a loop that executes a specified statement until the test condition evaluates to false.

ForInLoop

A for...in loop statement, as defined by the spec.

ForLoop

The for statement creates a loop that consists of three optional expressions.

ForOfLoop

A for...of loop statement, as defined by the spec.

WhileLoop

The while statement creates a loop that executes a specified statement as long as the test condition evaluates to true.

Enums

ForLoopInitializer

A ForLoop initializer, as defined by the spec.

IterableLoopInitializer

A for-in, for-of and for-await-of loop initializer.