Module boa_ast::visitor

ECMAScript Abstract Syntax Tree visitors.

This module contains visitors which can be used to inspect or modify AST nodes. This allows for fine-grained manipulation of ASTs for analysis, rewriting, or instrumentation.

Macros

define_visit 🔒

Creates the default visit function implementation for a particular type

define_visit_mut 🔒

Creates the default mutable visit function implementation for a particular type

node_ref 🔒

Generates the NodeRef and NodeMutRef enums from a list of variants.

Enums

NodeRef

A reference to a node visitable by a Visitor.

NodeRefMut

A mutable reference to a node visitable by a VisitorMut.

Traits

VisitWith

Denotes that a type may be visited, providing a method which allows a visitor to traverse its private fields.

Visitor

Represents an AST visitor.

VisitorMut

Represents an AST visitor which can modify AST content.