Trait boa_ast::visitor::VisitWith

pub trait VisitWith {
    // Required methods
    fn visit_with<'a, V>(&'a self, visitor: &mut V) -> ControlFlow<V::BreakTy>
       where V: Visitor<'a>;
    fn visit_with_mut<'a, V>(
        &'a mut self,
        visitor: &mut V
    ) -> ControlFlow<V::BreakTy>
       where V: VisitorMut<'a>;
}

Denotes that a type may be visited, providing a method which allows a visitor to traverse its private fields.

Required Methods

fn visit_with<'a, V>(&'a self, visitor: &mut V) -> ControlFlow<V::BreakTy>where V: Visitor<'a>,

Visit this node with the provided visitor.

fn visit_with_mut<'a, V>( &'a mut self, visitor: &mut V ) -> ControlFlow<V::BreakTy>where V: VisitorMut<'a>,

Visit this node with the provided visitor mutably, allowing the visitor to modify private fields.