Trait boa_ast::visitor::VisitorMut

pub trait VisitorMut<'ast>: Sized {
    type BreakTy;

    fn visit_statement_list_mut(
        &mut self,
        node: &'ast mut StatementList
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_statement_list_item_mut(
        &mut self,
        node: &'ast mut StatementListItem
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_statement_mut(
        &mut self,
        node: &'ast mut Statement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_declaration_mut(
        &mut self,
        node: &'ast mut Declaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_function_mut(
        &mut self,
        node: &'ast mut Function
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_generator_mut(
        &mut self,
        node: &'ast mut Generator
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_async_function_mut(
        &mut self,
        node: &'ast mut AsyncFunction
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_async_generator_mut(
        &mut self,
        node: &'ast mut AsyncGenerator
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_class_mut(
        &mut self,
        node: &'ast mut Class
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_lexical_declaration_mut(
        &mut self,
        node: &'ast mut LexicalDeclaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_block_mut(
        &mut self,
        node: &'ast mut Block
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_var_declaration_mut(
        &mut self,
        node: &'ast mut VarDeclaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_expression_mut(
        &mut self,
        node: &'ast mut Expression
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_if_mut(&mut self, node: &'ast mut If) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_do_while_loop_mut(
        &mut self,
        node: &'ast mut DoWhileLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_while_loop_mut(
        &mut self,
        node: &'ast mut WhileLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_for_loop_mut(
        &mut self,
        node: &'ast mut ForLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_for_in_loop_mut(
        &mut self,
        node: &'ast mut ForInLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_for_of_loop_mut(
        &mut self,
        node: &'ast mut ForOfLoop
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_switch_mut(
        &mut self,
        node: &'ast mut Switch
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_continue_mut(
        &mut self,
        node: &'ast mut Continue
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_break_mut(
        &mut self,
        node: &'ast mut Break
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_return_mut(
        &mut self,
        node: &'ast mut Return
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_labelled_mut(
        &mut self,
        node: &'ast mut Labelled
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_throw_mut(
        &mut self,
        node: &'ast mut Throw
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_try_mut(
        &mut self,
        node: &'ast mut Try
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_with_mut(
        &mut self,
        node: &'ast mut With
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_identifier_mut(
        &mut self,
        node: &'ast mut Identifier
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_formal_parameter_list_mut(
        &mut self,
        node: &'ast mut FormalParameterList
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_class_element_mut(
        &mut self,
        node: &'ast mut ClassElement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_private_name_mut(
        &mut self,
        node: &'ast mut PrivateName
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_variable_list_mut(
        &mut self,
        node: &'ast mut VariableList
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_variable_mut(
        &mut self,
        node: &'ast mut Variable
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_binding_mut(
        &mut self,
        node: &'ast mut Binding
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_pattern_mut(
        &mut self,
        node: &'ast mut Pattern
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_literal_mut(
        &mut self,
        node: &'ast mut Literal
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_array_literal_mut(
        &mut self,
        node: &'ast mut ArrayLiteral
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_object_literal_mut(
        &mut self,
        node: &'ast mut ObjectLiteral
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_spread_mut(
        &mut self,
        node: &'ast mut Spread
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_arrow_function_mut(
        &mut self,
        node: &'ast mut ArrowFunction
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_async_arrow_function_mut(
        &mut self,
        node: &'ast mut AsyncArrowFunction
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_template_literal_mut(
        &mut self,
        node: &'ast mut TemplateLiteral
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_property_access_mut(
        &mut self,
        node: &'ast mut PropertyAccess
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_new_mut(
        &mut self,
        node: &'ast mut New
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_call_mut(
        &mut self,
        node: &'ast mut Call
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_super_call_mut(
        &mut self,
        node: &'ast mut SuperCall
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_optional_mut(
        &mut self,
        node: &'ast mut Optional
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_tagged_template_mut(
        &mut self,
        node: &'ast mut TaggedTemplate
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_assign_mut(
        &mut self,
        node: &'ast mut Assign
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_unary_mut(
        &mut self,
        node: &'ast mut Unary
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_update_mut(
        &mut self,
        node: &'ast mut Update
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_binary_mut(
        &mut self,
        node: &'ast mut Binary
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_binary_in_private_mut(
        &mut self,
        node: &'ast mut BinaryInPrivate
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_conditional_mut(
        &mut self,
        node: &'ast mut Conditional
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_await_mut(
        &mut self,
        node: &'ast mut Await
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_yield_mut(
        &mut self,
        node: &'ast mut Yield
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_for_loop_initializer_mut(
        &mut self,
        node: &'ast mut ForLoopInitializer
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_iterable_loop_initializer_mut(
        &mut self,
        node: &'ast mut IterableLoopInitializer
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_case_mut(
        &mut self,
        node: &'ast mut Case
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_sym_mut(
        &mut self,
        node: &'ast mut Sym
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_labelled_item_mut(
        &mut self,
        node: &'ast mut LabelledItem
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_catch_mut(
        &mut self,
        node: &'ast mut Catch
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_finally_mut(
        &mut self,
        node: &'ast mut Finally
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_formal_parameter_mut(
        &mut self,
        node: &'ast mut FormalParameter
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_property_name_mut(
        &mut self,
        node: &'ast mut PropertyName
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_method_definition_mut(
        &mut self,
        node: &'ast mut MethodDefinition
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_object_pattern_mut(
        &mut self,
        node: &'ast mut ObjectPattern
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_array_pattern_mut(
        &mut self,
        node: &'ast mut ArrayPattern
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_property_definition_mut(
        &mut self,
        node: &'ast mut PropertyDefinition
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_template_element_mut(
        &mut self,
        node: &'ast mut TemplateElement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_simple_property_access_mut(
        &mut self,
        node: &'ast mut SimplePropertyAccess
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_private_property_access_mut(
        &mut self,
        node: &'ast mut PrivatePropertyAccess
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_super_property_access_mut(
        &mut self,
        node: &'ast mut SuperPropertyAccess
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_optional_operation_mut(
        &mut self,
        node: &'ast mut OptionalOperation
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_assign_target_mut(
        &mut self,
        node: &'ast mut AssignTarget
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_object_pattern_element_mut(
        &mut self,
        node: &'ast mut ObjectPatternElement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_array_pattern_element_mut(
        &mut self,
        node: &'ast mut ArrayPatternElement
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_property_access_field_mut(
        &mut self,
        node: &'ast mut PropertyAccessField
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_optional_operation_kind_mut(
        &mut self,
        node: &'ast mut OptionalOperationKind
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_module_item_list_mut(
        &mut self,
        node: &'ast mut ModuleItemList
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_module_item_mut(
        &mut self,
        node: &'ast mut ModuleItem
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_module_specifier_mut(
        &mut self,
        node: &'ast mut ModuleSpecifier
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_import_kind_mut(
        &mut self,
        node: &'ast mut ImportKind
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_import_declaration_mut(
        &mut self,
        node: &'ast mut ImportDeclaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_import_specifier_mut(
        &mut self,
        node: &'ast mut ImportSpecifier
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_re_export_kind_mut(
        &mut self,
        node: &'ast mut ReExportKind
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_export_declaration_mut(
        &mut self,
        node: &'ast mut ExportDeclaration
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit_export_specifier_mut(
        &mut self,
        node: &'ast mut ExportSpecifier
    ) -> ControlFlow<Self::BreakTy> { ... }
    fn visit<N: Into<NodeRefMut<'ast>>>(
        &mut self,
        node: N
    ) -> ControlFlow<Self::BreakTy> { ... }

}

Represents an AST visitor which can modify AST content.

This implementation is based largely on chalk’s visitor pattern.

Required Associated Types

type BreakTy

Type which will be propagated from the visitor if completing early.

Provided Methods

fn visit_statement_list_mut(
&mut self,
node: &'ast mut StatementList
) -> ControlFlow<Self::BreakTy>

Visits a StatementList with this visitor, mutably

fn visit_statement_list_item_mut(
&mut self,
node: &'ast mut StatementListItem
) -> ControlFlow<Self::BreakTy>

Visits a StatementListItem with this visitor, mutably

fn visit_statement_mut(
&mut self,
node: &'ast mut Statement
) -> ControlFlow<Self::BreakTy>

Visits a Statement with this visitor, mutably

fn visit_declaration_mut(
&mut self,
node: &'ast mut Declaration
) -> ControlFlow<Self::BreakTy>

Visits a Declaration with this visitor, mutably

fn visit_function_mut(
&mut self,
node: &'ast mut Function
) -> ControlFlow<Self::BreakTy>

Visits a Function with this visitor, mutably

fn visit_generator_mut(
&mut self,
node: &'ast mut Generator
) -> ControlFlow<Self::BreakTy>

Visits a Generator with this visitor, mutably

fn visit_async_function_mut(
&mut self,
node: &'ast mut AsyncFunction
) -> ControlFlow<Self::BreakTy>

Visits a AsyncFunction with this visitor, mutably

fn visit_async_generator_mut(
&mut self,
node: &'ast mut AsyncGenerator
) -> ControlFlow<Self::BreakTy>

Visits a AsyncGenerator with this visitor, mutably

fn visit_class_mut(
&mut self,
node: &'ast mut Class
) -> ControlFlow<Self::BreakTy>

Visits a Class with this visitor, mutably

fn visit_lexical_declaration_mut(
&mut self,
node: &'ast mut LexicalDeclaration
) -> ControlFlow<Self::BreakTy>

Visits a LexicalDeclaration with this visitor, mutably

fn visit_block_mut(
&mut self,
node: &'ast mut Block
) -> ControlFlow<Self::BreakTy>

Visits a Block with this visitor, mutably

fn visit_var_declaration_mut(
&mut self,
node: &'ast mut VarDeclaration
) -> ControlFlow<Self::BreakTy>

Visits a VarDeclaration with this visitor, mutably

fn visit_expression_mut(
&mut self,
node: &'ast mut Expression
) -> ControlFlow<Self::BreakTy>

Visits a Expression with this visitor, mutably

fn visit_if_mut(&mut self, node: &'ast mut If) -> ControlFlow<Self::BreakTy>

Visits a If with this visitor, mutably

fn visit_do_while_loop_mut(
&mut self,
node: &'ast mut DoWhileLoop
) -> ControlFlow<Self::BreakTy>

Visits a DoWhileLoop with this visitor, mutably

fn visit_while_loop_mut(
&mut self,
node: &'ast mut WhileLoop
) -> ControlFlow<Self::BreakTy>

Visits a WhileLoop with this visitor, mutably

fn visit_for_loop_mut(
&mut self,
node: &'ast mut ForLoop
) -> ControlFlow<Self::BreakTy>

Visits a ForLoop with this visitor, mutably

fn visit_for_in_loop_mut(
&mut self,
node: &'ast mut ForInLoop
) -> ControlFlow<Self::BreakTy>

Visits a ForInLoop with this visitor, mutably

fn visit_for_of_loop_mut(
&mut self,
node: &'ast mut ForOfLoop
) -> ControlFlow<Self::BreakTy>

Visits a ForOfLoop with this visitor, mutably

fn visit_switch_mut(
&mut self,
node: &'ast mut Switch
) -> ControlFlow<Self::BreakTy>

Visits a Switch with this visitor, mutably

fn visit_continue_mut(
&mut self,
node: &'ast mut Continue
) -> ControlFlow<Self::BreakTy>

Visits a Continue with this visitor, mutably

fn visit_break_mut(
&mut self,
node: &'ast mut Break
) -> ControlFlow<Self::BreakTy>

Visits a Break with this visitor, mutably

fn visit_return_mut(
&mut self,
node: &'ast mut Return
) -> ControlFlow<Self::BreakTy>

Visits a Return with this visitor, mutably

fn visit_labelled_mut(
&mut self,
node: &'ast mut Labelled
) -> ControlFlow<Self::BreakTy>

Visits a Labelled with this visitor, mutably

fn visit_throw_mut(
&mut self,
node: &'ast mut Throw
) -> ControlFlow<Self::BreakTy>

Visits a Throw with this visitor, mutably

fn visit_try_mut(&mut self, node: &'ast mut Try) -> ControlFlow<Self::BreakTy>

Visits a Try with this visitor, mutably

fn visit_with_mut(&mut self, node: &'ast mut With) -> ControlFlow<Self::BreakTy>

Visits a With with this visitor, mutably

fn visit_identifier_mut(
&mut self,
node: &'ast mut Identifier
) -> ControlFlow<Self::BreakTy>

Visits a Identifier with this visitor, mutably

fn visit_formal_parameter_list_mut(
&mut self,
node: &'ast mut FormalParameterList
) -> ControlFlow<Self::BreakTy>

Visits a FormalParameterList with this visitor, mutably

fn visit_class_element_mut(
&mut self,
node: &'ast mut ClassElement
) -> ControlFlow<Self::BreakTy>

Visits a ClassElement with this visitor, mutably

fn visit_private_name_mut(
&mut self,
node: &'ast mut PrivateName
) -> ControlFlow<Self::BreakTy>

Visits a PrivateName with this visitor, mutably

fn visit_variable_list_mut(
&mut self,
node: &'ast mut VariableList
) -> ControlFlow<Self::BreakTy>

Visits a VariableList with this visitor, mutably

fn visit_variable_mut(
&mut self,
node: &'ast mut Variable
) -> ControlFlow<Self::BreakTy>

Visits a Variable with this visitor, mutably

fn visit_binding_mut(
&mut self,
node: &'ast mut Binding
) -> ControlFlow<Self::BreakTy>

Visits a Binding with this visitor, mutably

fn visit_pattern_mut(
&mut self,
node: &'ast mut Pattern
) -> ControlFlow<Self::BreakTy>

Visits a Pattern with this visitor, mutably

fn visit_literal_mut(
&mut self,
node: &'ast mut Literal
) -> ControlFlow<Self::BreakTy>

Visits a Literal with this visitor, mutably

fn visit_array_literal_mut(
&mut self,
node: &'ast mut ArrayLiteral
) -> ControlFlow<Self::BreakTy>

Visits a ArrayLiteral with this visitor, mutably

fn visit_object_literal_mut(
&mut self,
node: &'ast mut ObjectLiteral
) -> ControlFlow<Self::BreakTy>

Visits a ObjectLiteral with this visitor, mutably

fn visit_spread_mut(
&mut self,
node: &'ast mut Spread
) -> ControlFlow<Self::BreakTy>

Visits a Spread with this visitor, mutably

fn visit_arrow_function_mut(
&mut self,
node: &'ast mut ArrowFunction
) -> ControlFlow<Self::BreakTy>

Visits a ArrowFunction with this visitor, mutably

fn visit_async_arrow_function_mut(
&mut self,
node: &'ast mut AsyncArrowFunction
) -> ControlFlow<Self::BreakTy>

Visits a AsyncArrowFunction with this visitor, mutably

fn visit_template_literal_mut(
&mut self,
node: &'ast mut TemplateLiteral
) -> ControlFlow<Self::BreakTy>

Visits a TemplateLiteral with this visitor, mutably

fn visit_property_access_mut(
&mut self,
node: &'ast mut PropertyAccess
) -> ControlFlow<Self::BreakTy>

Visits a PropertyAccess with this visitor, mutably

fn visit_new_mut(&mut self, node: &'ast mut New) -> ControlFlow<Self::BreakTy>

Visits a New with this visitor, mutably

fn visit_call_mut(&mut self, node: &'ast mut Call) -> ControlFlow<Self::BreakTy>

Visits a Call with this visitor, mutably

fn visit_super_call_mut(
&mut self,
node: &'ast mut SuperCall
) -> ControlFlow<Self::BreakTy>

Visits a SuperCall with this visitor, mutably

fn visit_optional_mut(
&mut self,
node: &'ast mut Optional
) -> ControlFlow<Self::BreakTy>

Visits a Optional with this visitor, mutably

fn visit_tagged_template_mut(
&mut self,
node: &'ast mut TaggedTemplate
) -> ControlFlow<Self::BreakTy>

Visits a TaggedTemplate with this visitor, mutably

fn visit_assign_mut(
&mut self,
node: &'ast mut Assign
) -> ControlFlow<Self::BreakTy>

Visits a Assign with this visitor, mutably

fn visit_unary_mut(
&mut self,
node: &'ast mut Unary
) -> ControlFlow<Self::BreakTy>

Visits a Unary with this visitor, mutably

fn visit_update_mut(
&mut self,
node: &'ast mut Update
) -> ControlFlow<Self::BreakTy>

Visits a Update with this visitor, mutably

fn visit_binary_mut(
&mut self,
node: &'ast mut Binary
) -> ControlFlow<Self::BreakTy>

Visits a Binary with this visitor, mutably

fn visit_binary_in_private_mut(
&mut self,
node: &'ast mut BinaryInPrivate
) -> ControlFlow<Self::BreakTy>

Visits a BinaryInPrivate with this visitor, mutably

fn visit_conditional_mut(
&mut self,
node: &'ast mut Conditional
) -> ControlFlow<Self::BreakTy>

Visits a Conditional with this visitor, mutably

fn visit_await_mut(
&mut self,
node: &'ast mut Await
) -> ControlFlow<Self::BreakTy>

Visits a Await with this visitor, mutably

fn visit_yield_mut(
&mut self,
node: &'ast mut Yield
) -> ControlFlow<Self::BreakTy>

Visits a Yield with this visitor, mutably

fn visit_for_loop_initializer_mut(
&mut self,
node: &'ast mut ForLoopInitializer
) -> ControlFlow<Self::BreakTy>

Visits a ForLoopInitializer with this visitor, mutably

fn visit_iterable_loop_initializer_mut(
&mut self,
node: &'ast mut IterableLoopInitializer
) -> ControlFlow<Self::BreakTy>

Visits a IterableLoopInitializer with this visitor, mutably

fn visit_case_mut(&mut self, node: &'ast mut Case) -> ControlFlow<Self::BreakTy>

Visits a Case with this visitor, mutably

fn visit_sym_mut(&mut self, node: &'ast mut Sym) -> ControlFlow<Self::BreakTy>

Visits a Sym with this visitor, mutably

fn visit_labelled_item_mut(
&mut self,
node: &'ast mut LabelledItem
) -> ControlFlow<Self::BreakTy>

Visits a LabelledItem with this visitor, mutably

fn visit_catch_mut(
&mut self,
node: &'ast mut Catch
) -> ControlFlow<Self::BreakTy>

Visits a Catch with this visitor, mutably

fn visit_finally_mut(
&mut self,
node: &'ast mut Finally
) -> ControlFlow<Self::BreakTy>

Visits a Finally with this visitor, mutably

fn visit_formal_parameter_mut(
&mut self,
node: &'ast mut FormalParameter
) -> ControlFlow<Self::BreakTy>

Visits a FormalParameter with this visitor, mutably

fn visit_property_name_mut(
&mut self,
node: &'ast mut PropertyName
) -> ControlFlow<Self::BreakTy>

Visits a PropertyName with this visitor, mutably

fn visit_method_definition_mut(
&mut self,
node: &'ast mut MethodDefinition
) -> ControlFlow<Self::BreakTy>

Visits a MethodDefinition with this visitor, mutably

fn visit_object_pattern_mut(
&mut self,
node: &'ast mut ObjectPattern
) -> ControlFlow<Self::BreakTy>

Visits a ObjectPattern with this visitor, mutably

fn visit_array_pattern_mut(
&mut self,
node: &'ast mut ArrayPattern
) -> ControlFlow<Self::BreakTy>

Visits a ArrayPattern with this visitor, mutably

fn visit_property_definition_mut(
&mut self,
node: &'ast mut PropertyDefinition
) -> ControlFlow<Self::BreakTy>

Visits a PropertyDefinition with this visitor, mutably

fn visit_template_element_mut(
&mut self,
node: &'ast mut TemplateElement
) -> ControlFlow<Self::BreakTy>

Visits a TemplateElement with this visitor, mutably

fn visit_simple_property_access_mut(
&mut self,
node: &'ast mut SimplePropertyAccess
) -> ControlFlow<Self::BreakTy>

Visits a SimplePropertyAccess with this visitor, mutably

fn visit_private_property_access_mut(
&mut self,
node: &'ast mut PrivatePropertyAccess
) -> ControlFlow<Self::BreakTy>

Visits a PrivatePropertyAccess with this visitor, mutably

fn visit_super_property_access_mut(
&mut self,
node: &'ast mut SuperPropertyAccess
) -> ControlFlow<Self::BreakTy>

Visits a SuperPropertyAccess with this visitor, mutably

fn visit_optional_operation_mut(
&mut self,
node: &'ast mut OptionalOperation
) -> ControlFlow<Self::BreakTy>

Visits a OptionalOperation with this visitor, mutably

fn visit_assign_target_mut(
&mut self,
node: &'ast mut AssignTarget
) -> ControlFlow<Self::BreakTy>

Visits a AssignTarget with this visitor, mutably

fn visit_object_pattern_element_mut(
&mut self,
node: &'ast mut ObjectPatternElement
) -> ControlFlow<Self::BreakTy>

Visits a ObjectPatternElement with this visitor, mutably

fn visit_array_pattern_element_mut(
&mut self,
node: &'ast mut ArrayPatternElement
) -> ControlFlow<Self::BreakTy>

Visits a ArrayPatternElement with this visitor, mutably

fn visit_property_access_field_mut(
&mut self,
node: &'ast mut PropertyAccessField
) -> ControlFlow<Self::BreakTy>

Visits a PropertyAccessField with this visitor, mutably

fn visit_optional_operation_kind_mut(
&mut self,
node: &'ast mut OptionalOperationKind
) -> ControlFlow<Self::BreakTy>

Visits a OptionalOperationKind with this visitor, mutably

fn visit_module_item_list_mut(
&mut self,
node: &'ast mut ModuleItemList
) -> ControlFlow<Self::BreakTy>

Visits a ModuleItemList with this visitor, mutably

fn visit_module_item_mut(
&mut self,
node: &'ast mut ModuleItem
) -> ControlFlow<Self::BreakTy>

Visits a ModuleItem with this visitor, mutably

fn visit_module_specifier_mut(
&mut self,
node: &'ast mut ModuleSpecifier
) -> ControlFlow<Self::BreakTy>

Visits a ModuleSpecifier with this visitor, mutably

fn visit_import_kind_mut(
&mut self,
node: &'ast mut ImportKind
) -> ControlFlow<Self::BreakTy>

Visits a ImportKind with this visitor, mutably

fn visit_import_declaration_mut(
&mut self,
node: &'ast mut ImportDeclaration
) -> ControlFlow<Self::BreakTy>

Visits a ImportDeclaration with this visitor, mutably

fn visit_import_specifier_mut(
&mut self,
node: &'ast mut ImportSpecifier
) -> ControlFlow<Self::BreakTy>

Visits a ImportSpecifier with this visitor, mutably

fn visit_re_export_kind_mut(
&mut self,
node: &'ast mut ReExportKind
) -> ControlFlow<Self::BreakTy>

Visits a ReExportKind with this visitor, mutably

fn visit_export_declaration_mut(
&mut self,
node: &'ast mut ExportDeclaration
) -> ControlFlow<Self::BreakTy>

Visits a ExportDeclaration with this visitor, mutably

fn visit_export_specifier_mut(
&mut self,
node: &'ast mut ExportSpecifier
) -> ControlFlow<Self::BreakTy>

Visits a ExportSpecifier with this visitor, mutably

fn visit<N: Into<NodeRefMut<'ast>>>(
&mut self,
node: N
) -> ControlFlow<Self::BreakTy>

Generic entry point for a node that is visitable by a VisitorMut.

This is usually used for generic functions that need to visit an unnamed AST node.